+++ circuits/src/generation/ctl_test/debug_trace_print.rs
@@ -13,6 +13,7 @@ use std::{
 use assembler::encoder::encode_asm_from_json_file;
 use executor::{load_tx::init_tape, Process};
 use plonky2::hash::hash_types::RichField;
+use core::vm::transaction::init_tx_context;
 
 #[allow(unused)]
 fn get_looking_looked_info<
@@ -167,7 +168,7 @@ pub fn get_exec_trace(file_name: String, call_data: Option<Vec<GoldilocksField>>
     process.addr_storage = Address::default();
     if let Some(calldata) = call_data {
         process.tp = GoldilocksField::ZERO;
-        init_tape(&mut process, calldata, Address::default(), Address::default(), Address::default());
+        init_tape(&mut process, calldata, Address::default(), Address::default(), Address::default(), &init_tx_context());
     }
     let _ = process.execute(
         &mut program,
