+++ circuits/src/test_utils.rs
@@ -8,6 +8,8 @@ use plonky2_util::log2_strict;
 
 use crate::stark::{constraint_consumer::ConstraintConsumer, vars::StarkEvaluationVars};
 use core::merkle_tree::tree::AccountTree;
+use core::vm::transaction::init_tx_context;
+
 pub fn test_stark_with_asm_path<Row, const COL_NUM: usize, E, H>(
     path: String,
     get_trace_rows: fn(Trace) -> Vec<Row>,
@@ -44,7 +46,7 @@ pub fn test_stark_with_asm_path<Row, const COL_NUM: usize, E, H>(
 
     if let Some(calldata) = call_data {
         process.tp = GoldilocksField::ZERO;
-        init_tape(&mut process, calldata, Address::default(), Address::default(), Address::default());
+        init_tape(&mut process, calldata, Address::default(), Address::default(), Address::default(), &init_tx_context());
     }
 
     let _ = process.execute(
