+++ core/src/vm/mod.rs
@@ -5,3 +5,4 @@ pub mod memory;
 pub mod opcodes;
 pub mod operands;
 pub mod vm_state;
+pub mod transaction;
